app/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pyjwt
Jinja2
python-multipart